Moved Bandwidth Result structures and comparison statistics to a separate file
