+++ tests/tt_metal/tt_metal/perf_microbenchmark/CMakeLists.txt
@@ -67,6 +67,7 @@ foreach(TEST_SRC ${PERF_MICROBENCH_TESTS_SRCS})
             PRIVATE
                 routing/tt_fabric_test_common_types.cpp
                 routing/tt_fabric_test_config.cpp
+                routing/tt_fabric_test_results.cpp
                 routing/tt_fabric_test_context.cpp
         )
         target_link_libraries(${TEST_TARGET} PRIVATE yaml-cpp::yaml-cpp)
